+++ src/constants.ts
@@ -1,0 +1,8 @@
+export const THEME_PLAYGROUND = 'storybook-addon-theme-playground';
+
+export const THEME_PLAYGROUND_PANEL = 'storybook-addon-theme-playground/panel';
+export const THEME_PLAYGROUND_STATE = 'storybook-addon-theme-playground/state';
+
+export const THEME_PLAYGROUND_RESET = 'storybook-addon-theme-playground/reset';
+export const THEME_PLAYGROUND_UPDATE =
+  'storybook-addon-theme-playground/update';