+++ src/events/index.ts
@@ -1,9 +1,0 @@
-export default {
-  receiveOptions: 'storybook-addon-theme-playground/receiveOptions',
-  resetOptions: 'storybook-addon-theme-playground/resetOptions',
-  reset: 'storybook-addon-theme-playground/reset',
-  setTheme: 'storybook-addon-theme-playground/setTheme',
-  updateTheme: 'storybook-addon-theme-playground/updateTheme',
-  setOverrides: 'storybook-addon-theme-playground/setOverrides',
-  setConfig: 'storybook-addon-theme-playground/setConfig'
-};