+++ src/helper/buildThemeComponents.ts
@@ -1,5 +1,22 @@
-import { is } from './';
+import { is } from '.';
 import { OverridesProps } from '../types';
+
+export const getThemeComponents = (
+  theme: any | { [key: string]: any }[],
+  overrides: OverridesProps
+) => {
+  const components: { [key: string]: any } = {};
+
+  if (Array.isArray(theme)) {
+    theme.forEach((item) => {
+      components[item.name] = buildThemeComponents(item.theme, overrides);
+    });
+  } else {
+    components['Default Theme'] = buildThemeComponents(theme, overrides);
+  }
+
+  return components;
+};
 
 const buildThemeComponents = (
   theme: any,
@@ -8,61 +25,60 @@ const buildThemeComponents = (
 ) => {
   const themeComponents = Object.keys(theme).reduce((acc, key) => {
     const value = theme[key];
-    const path = [...arr, key];
-    const pathString = path.join('.');
+    const pathArray = [...arr, key];
+    const path = pathArray.join('.');
 
-    // Return if component is hidden
-    if (overrides[pathString] && overrides[pathString].hidden) {
+    if (overrides[path] && overrides[path].hidden) {
       return;
     }
 
-    if (overrides[pathString] && overrides[pathString].type) {
+    if (overrides[path] && overrides[path].type) {
       return {
         ...acc,
-        [pathString]: { type: overrides[pathString].type, value },
+        [path]: { type: overrides[path].type, value },
       };
     }
 
     if (Array.isArray(value)) {
       return {
         ...acc,
-        ...buildThemeComponents(value, overrides, path),
+        ...buildThemeComponents(value, overrides, pathArray),
       };
     }
 
     if (is.object(value)) {
       if (is.shorthand(value)) {
-        return { ...acc, [pathString]: { type: 'shorthand', value } };
+        return { ...acc, [path]: { type: 'shorthand', value } };
       }
 
       return {
         ...acc,
-        ...buildThemeComponents(value, overrides, path),
+        ...buildThemeComponents(value, overrides, pathArray),
       };
     }
 
     if (is.boolean(value)) {
-      return { ...acc, [pathString]: { type: 'switch', value } };
+      return { ...acc, [path]: { type: 'switch', value } };
     }
 
     if (typeof value === 'number') {
-      return { ...acc, [pathString]: { type: 'counter', value } };
+      return { ...acc, [path]: { type: 'number', value } };
     }
 
     if (typeof value === 'string') {
       if (is.color(value, key)) {
-        return { ...acc, [pathString]: { type: 'colorpicker', value } };
+        return { ...acc, [path]: { type: 'colorpicker', value } };
       }
 
       if (is.unit(value)) {
-        return { ...acc, [pathString]: { type: 'range', value } };
+        return { ...acc, [path]: { type: 'range', value } };
       }
 
       if (is.text(value)) {
-        return { ...acc, [pathString]: { type: 'textarea', value } };
+        return { ...acc, [path]: { type: 'textarea', value } };
       }
 
-      return { ...acc, [pathString]: { type: 'input', value } };
+      return { ...acc, [path]: { type: 'input', value } };
     }
   }, {});