+++ src/helper/stripUnit.ts
@@ -1,7 +1,0 @@
-export const stripUnit = (value: string | number): any => {
-  if (typeof value !== 'string') return [value, undefined];
-  const matchedValue = value.match(/^([+-]?(?:\d+|\d*\.\d+))([a-z]*|%)$/);
-
-  if (matchedValue) return [parseFloat(value), matchedValue[2]];
-  return [value, undefined];
-};