+++ src/withThemePlayground.tsx
@@ -1,9 +1,20 @@
-/* eslint-disable react/display-name */
-import React from 'react';
-import addons from '@storybook/addons';
+import React, { useMemo } from 'react';
+import addons, {
+  makeDecorator,
+  useCallback,
+  useEffect,
+  useState,
+} from '@storybook/addons';
+import { useAddonState } from '@storybook/client-api';
 
-import { ConfigProps, OverridesProps } from './types';
-import events from './events';
+import { ConfigProps, OverridesProps, PanelState } from './types';
+import {
+  THEME_PLAYGROUND_RESET,
+  THEME_PLAYGROUND_STATE,
+  THEME_PLAYGROUND_UPDATE,
+} from './constants';
+
+import { getThemeComponents } from './helper/buildThemeComponents';
 
 export interface ThemePlaygroundProps {
   theme: any;
@@ -12,65 +23,76 @@ export interface ThemePlaygroundProps {
   config?: ConfigProps;
 }
 
-type ThemePlaygroundState = any;
-
-export class WithThemePlayground extends React.Component<
-  ThemePlaygroundProps,
-  ThemePlaygroundState
-> {
-  state = Array.isArray(this.props.theme)
-    ? this.props.theme[0].theme
-    : this.props.theme;
-
-  channel = addons.getChannel();
-
-  handleReset = () => {
-    const { theme, config, overrides } = this.props;
-    return this.channel.emit(events.resetOptions, { theme, config, overrides });
-  };
-
-  componentDidMount() {
-    const { theme, config, overrides } = this.props;
-
-    this.channel.on(events.updateTheme, (t) => this.setState(t));
-    this.channel.on(events.reset, this.handleReset);
-
-    this.channel.emit(events.receiveOptions, { theme, config, overrides });
-  }
-
-  componentWillUnmount() {
-    this.channel.removeListener(events.updateTheme, (t) => this.setState(t));
-    this.channel.removeListener(events.reset, this.handleReset);
-  }
-
-  render() {
-    const ThemeProvider = this.props.provider;
-
-    return (
-      <ThemeProvider theme={this.state}>{this.props.children}</ThemeProvider>
-    );
-  }
-}
-
-export const withThemePlayground = (props: ThemePlaygroundProps) => {
-  if (!props.provider) {
-    throw Error(
-      'Missing ThemeProvider in withThemePlayground decorator options.'
-    );
-  }
-
-  if (!props.theme) {
-    throw Error('Missing theme key in withThemePlayground decorator options.');
-  }
-
-  return (storyFn) => (
-    <WithThemePlayground
-      theme={props.theme}
-      provider={props.provider}
-      config={props.config}
-      overrides={props.overrides}
-    >
-      {storyFn()}
-    </WithThemePlayground>
-  );
+const getThemeArray = (theme) => {
+  return Array.isArray(theme) ? theme : [{ name: 'Default Theme', theme }];
 };
+
+export default makeDecorator({
+  name: 'withThemePlayground',
+  parameterName: 'themePlayground',
+  skipIfNoParametersOrOptions: true,
+  wrapper: (storyFn, context, { options }) => {
+    const { theme, overrides, config, provider } = options;
+
+    const channel = addons.getChannel();
+
+    if (!provider) {
+      throw Error(
+        'Missing ThemeProvider in withThemePlayground decorator options.'
+      );
+    }
+
+    if (!theme) {
+      throw Error(
+        'Missing theme key in withThemePlayground decorator options.'
+      );
+    }
+
+    const initialState: PanelState = {
+      selected: 0,
+      theme: getThemeArray(theme),
+      themeComponents: useMemo(() => getThemeComponents(theme, overrides), []),
+      overrides,
+      config: {
+        labelFormat: 'startCase',
+        debounce: true,
+        debounceRate: 500,
+        showCode: true,
+        ...config,
+      },
+    };
+
+    const [state, setState] = useAddonState<PanelState>(
+      THEME_PLAYGROUND_STATE,
+      { ...initialState }
+    );
+
+    const [providerTheme, setProviderTheme] = useState(
+      state.theme[state.selected].theme
+    );
+
+    const handleReset = useCallback((i) => {
+      setState({
+        ...state,
+        selected: i,
+        theme: getThemeArray(theme),
+        themeComponents: getThemeComponents(theme, overrides),
+      });
+    }, []);
+
+    const handleUpdate = (newTheme) => setProviderTheme(newTheme);
+
+    useEffect(() => {
+      channel.on(THEME_PLAYGROUND_RESET, handleReset);
+      channel.on(THEME_PLAYGROUND_UPDATE, handleUpdate);
+      return () => {
+        channel.off(THEME_PLAYGROUND_RESET, handleReset);
+        channel.off(THEME_PLAYGROUND_UPDATE, handleUpdate);
+      };
+    }, []);
+
+    const Provider = provider;
+
+    return <Provider theme={providerTheme}>{storyFn(context)}</Provider>;
+  },
+});